Fixes #7994. Clears the bookmark in some cases when exiting playback.
If the user has PlaybackExitPrompt set to 2 ("Save position and
exit"), then the bookmark should be set any time playback exits, even
at the natural end of playback.  In this case, any existing bookmark
is explicitly cleared, since a bookmark at the end of a recording is
essentially equivalent to a bookmark at the very beginning of the
recording which is equivalent to no bookmark at all.