Cutlist editor: Add new theme strings for editing.
New strings are available to the theme for displaying cutlist-adjusted
time values in the editor:

reltimedisplay: "relative" version of timedisplay.  This is the
effective timedisplay after the cutlist is applied.

rellengthdisplay: the effective length of the program after the
cutlist is applied.

fulltimedisplay: a convenience string that is constructed as
"timedisplay (reltimedisplay of rellengthdisplay)".  Example:
    02:11.24 (00:13 of 23:59)