Take the cutlist into account when seeking during playback.
Cut regions are honored when calculating exactly where to jump.

Temporarily, the "INFOWITHCUTLIST" key binding is added for the
Arbitrary Seek functionality.  By default, "INFO" will jump to an
absolute (non cutlist adjusted) point in the recording, whereas
"INFOWITHCUTLIST" with do a cutlist-adjusted jump.  This can be
removed when we work out how to allow the playback progress bar to
convey cutlist information.

Refs #8631.