DB update for RTP. This adds an iptv_channel table.
In the existing IPTV recorder we need to read in a text file on
startup for the RTSP/RTP/UDP urls, this adds a table for those
so we can keep this data in the database. This requires a new
column in channel so we can link to this table. I've also defined
several URL types since an RTP stream require up to three URLs.

I'm adding this now so I can more easily switch between branches
and so others can test the rtp and ceton branches without needing
to manually edit the database.