Add assorted authenticated methods.
This adds methods for setting ratings and favorite status against
movies, as well as pulling lists of movies for which those have been
set. This adds a bypass in the cache to not operate on anything with a
lifetime of '0'.
(cherry picked from commit fb7c7cd3dfd4692e6c5acc61f0d2fd927b04d50e)