Fix Live TV channel changing when using Input Groups.
When trying to change channels (either directly or through the program
guide), an idle tuner in the same input group as the current tuner
would not be considered available.

This is fixed by adding the excluded_cardids argument to
RemoteRequestFreeRecorderList() and
RemoteRequestFreeRecorderFromList().

Doing this the "right" way would require a protocol change for the
GET_FREE_RECORDER_LIST command.  This is better done closer to the end
of the 0.26 release cycle, so the current implementation is less
efficient and makes multiple queries.

There is also a lot of vector<> code here that could stand to be
converted to QVector<>.