Add rsOtherTuning and rsOtherRecording recording statuses.
rsOtherShowing is a display only recording status used by mythfrontend
to indicate a program will record or is recording on another related
channel.  There are two long-standing problems with it.  One, it
doesn't distinguish between will record and is recording.  Two, it is
treated as a non-recording type.  Both problems force the user to find
the "other" channel in the EPG or elsewhere in order to see the
complete status and/or maniuplate the recording.  This change
addresses both problems.

Add new rsOtherTuning and rsOtherRecording recording statuses.  These
statuses are analogous to the existing rsTuning and rsRecording
statuses and indicates that a recording is active.  rsOtherShowing is
now analogous to rsWillRecord and indicates future activity.

Treat all of the rsOther* status like their analogs.  This makes it
possible for actions like stoping an active recording or adding an
override without having to navigate to the other channel.

This change also removes oneChanid parameter from the ProgramInfo
constructor and from LoadFromProgram since it is no longer needed.