Live TV: Disable the oldfile ringbuffer optimization for program transitions.
Refs #10490.  Using the LiveTVChain in the frontend ringbuffer to
control the oldfile flag in the backend ringbuffer is just too tricky
to get right and too easy to subsequently break, so we disable that
mechanism completely.  The consequence is that the backend ringbuffer
may cause a 0.3 second pause waiting for more data at true EOF (but
buffering on the frontend will probably hide that).  If the oldfile
optimization during live TV is really needed, it's probably better for
the backend recorded to control it.