Rework RAOP protocol handling
This commit fixes several issues in the RAOP protocol initialisation. Major ones being:
*iTunes compatibility. Playback from iTunes client now works (tested with both Mac and Windows version).
*Commit 37385baff83eea77116d13e22be96b74cfde2cec broke playback rather significantly, the RAOP server would never received packets being re-issued resulting in audio drop-outs

The core issue for iTunes not working previously was how the response packets were formed. The headers/content were inversed. This worked fine for iPhone/iPad but iTunes didn't like it.

This commit doesn't fix some audio-sync issues and occasional glitches in playback. This will be done in a future commit

[cherry-picked from 640ddc3e0419bf1396ceabf23e38fc3c1a56a200]