Add ability to override Bonjour's service name.
This is a work around a more complex issue.
By default, Bonjour would automatically detect the computer name and use it to advertise the server.

From Bonjour man-page:
"If non-NULL, specifies the service name to be registered. Most applications will not specify a name, in which case the computer name is used (this name is communicated to the client via the callback). If a name is specified, it must be 1-63 bytes of UTF-8 text. If the name is longer than 63 bytes it will be automatically truncated to a legal length"

However, for some people with broken DNS configuration, they can't resolve that name or it resolves improperly.
So add a BonjourHostname setting to override the name used by Bonjour.

Use it as an extra argument: -O BonjourHostname=mycomputer.

[cherry-picked from 48d84958a93e94e97591f0f239902fecfca8607f]