Add full IPv6 support to RAOP and fix Bonjour discovery of RAOP service.
Previously, what hostname Bonjour used to advertise a service was a bit of a mystery. If IPv6 was available , it could very well advertise an hostname resolving as IPv6 or use a link-local address. This cause the service to be visible on client (here iTunes), but it wouldn't have worked.
We know listen on all local interfaces, IPv6 and IPv4.
Also, do not advertise RAOP via Bonjour, if listening on the local interfaces failed