Added 5 min idle shutdown
When the last client (other than the logserver itself) disconnects, a 5min
timer is started.  If no clients connect during that time, the server will
shut down.  As it is automatically started up by a client that needs it, this
will keep it from running indefinitely when not needed, and also will allow
it to restart cleanly over upgrades.