Make sure UPnP autoconfiguration allows backends time to respond.
According to the UPnP spec the backends have up to the timeout value to respond
but we don't wait that long for all backends to respond so long as one backend
has responded.

Our backends try to respond ASAP so this isn't as bad as doing this with other
UPnP devices, but there is a race condition which results in us connecting to
the first backend to respond instead of popping up the chooser when there is
more than one backend on the network.

This also resends the search request every 250 ms. When testing this I ran into
some problems due to lost packets. When the broadcast search packet doesn't make
it onto the network we obviously never get a response. This just resends that
packet every 250 ms until only one second remains for the backends to respond.
This increases the odds that we get a response in the presence of an unreliable
network (in my case WiFi with Ubuntu Precise & bug 836250).