Add command string parser to break up commands for direct execution.
This adds a parser for the UNIX version of MythSystem, to break commands
apart at white space characters, obeying quoting and escaping where
appropriate, to allow commands passed as a single string to be executed
directly by execv, rather than be passed through the Bourne shell.
Note, this does not include support for any form of piping, redirects,
or backgrounding.  This includes an ABI bump.

Refs #10860