Enable command splitter for anything not explicitly flagging kMSRunShell
This alters the MythSystem(QString, uint) constructor to automatically
parse any commands passed through it that do not explicitly set the
kMSRunShell flag.  A quick run through the code does not turn up any
instances where this is currently the case, so the code should remain
effectively unused for now.  Individual MythSystem users and the block
of myth_system() calls will need to be tested individually with this
path to debug the parser.

Ref #10860