Add recording rule templates.
Recording rule templates are used to initialize new recording rules.
Templates can also be used to modify existing rules.  The "Default"
template replaces several, individual settings previously available in
the Setup wizard.  Those settings are automatically propagated to the
Default template.

Templates can be added, edited and deleted from within the Recording
Rules screen.  To add a new template, choose MENU / "New Template" and
then enter the name of the new template.  To edit a template,
highlight it and then choose ENTER or EDIT to bring up the schedule
editor.  To delete a template, highlight it and then choose DELETE or
edit it and change the type to "Delete this recording rule template."

When a regular rule is created, MythTV first tries to use the template
whose name matches the category of the program.  Failing that, it
tries to use the template whose name matches the category type of the
program.  If no template is found which matches the category or
category type, the Default template is used.

For example, if a user frequently records baseball games, he can
create a "Baseball" template which automatically sets the end-late
option to allow for extra innings.  Likewise, a user can create a
"Series" template which automatically sets the duplicate checking
method and the episode limit.

When a search, manual or template rule is created, the Default
template is always used.

The names of template need not match program categories or category
types.  They can simply be available to apply to other rules.  To
apply a different template to an existing rule, including a template
rule, begin editing the rule and then press MENU / "Use Template."
For example, a user can create an "Archive" template which changes the
recording group to "Archive" and turns off auto-expiry.