Update Python bindings to use the new config.xml format.
This creates a new DatabaseConfig class to manage access credentials for
the database, moving it out of DBCache.  DBCache (and MythDB) will still
continue to take the same arguments it did previously.  Both the old and
new format are supported, and old format files will be rewritten upon
successful connection to the database.  Files are now properly
pretty-printed when created.

This also converts everything over to absolute importing.

Fixes #10692.