Fix order of startup for logging in logserver
So that the logserver's logs are in the same order as other programs, we need
to half-start LoggingThread, then half-start LogServerThread (so the ZeroMQ
queues are available to attach to).  After that, both threads are welcome to
continue on their merry way.

Also, tweaked the SIGHUP handler log message to match the others.