Split the logserver into two threads
It seems under stress, the heartbeats get lost, causing all of the clients to
reconnect.  To combat this, I have moved the reception into its own thread,
and the logging into another.  In doing this, I have created two separate
ZeroMQ contexts as well, and changed the polling interval for the TCP reception
to 1ms rather than the default 10ms.