Speed up schema upgrades.
The old schema upgrade didn't lock before checking the schema version, so two processes
try to upgrade the same schema. To try to avoid this happening in practice we would sleep
for a few seconds whenever the schema didn't match what we expected. This was safe since
the actual schema upgrade used a lock, but it was unnecessarily slow.

Now we just lock before we check the schema version. If the schema is locked we wait,
if not we proceed.