Remove path names from HLS playlist entries.
The spec does not require path names, so remove them so that all files
are found relative to the playlist itself.  The path was only originally
supplied in the playlist due to the fact that we had to use URIs of
the form /Content/GetFile?StorageGroup=blah&FileName=bleh.  Now that
we can use /StorageGroup/blah/bleh we no longer need to use paths in
the files since clients can follow the new URI format.