Subtitles: Make the subtitle zoom factor into a live setting.
OSDCC708TextZoom is removed from the frontend OSD setup screen and
added as a "live setting" to the Subtitles menu of the playback OSD.
This brings up an adjustment dialog similar to adjusting audio sync or
timestretch.

The menu item is only offered when subtitles are currently enabled.

For text and CEA-708 subtitles, adjustments take effect immediately.
For CEA-608, the adjustment takes effect only when the next caption is
displayed, due to the way the 608 reader is structured.

For now, only the Subtitles menu entry is provided, but if needed, it
would be practical to add a wider set of actions and keybindings to
match the implementation of audio sync and timestretch adjustment.