Fix seeking in HLS playback..
Can finally seek properly ; and pretty accurate and fast at that !
Seeking in Live stream is enabled. This can have some adverse effects if attempting to seek to fast toward the end and a EOF may be detected. However, this allows starting the HLS stream generated by myth and seek into it properly.
Bookmarks work.

Also improve retry when the download of a segments failed. After 3 attempts, the segments is simply dropped. I figured this is better than dropping to a lower quality stream because 1-it may not be any better on the other stream, and 2-FFmpeg H264 decoder will crash if there's a change of resolution