Add a playback OSD menu option to cancel an in-progress playlist.
Previously, the only way to stop playlist playback was to execute a
jump point, as EXIT simply moves to the next playlist item.  Now, the
playback OSD menu, "Playback" submenu, has a "Cancel Playlist" entry,
with the effect that when playback ends, it returns to the Watch
Recordings screen instead of the next playlist entry.

Note: This adds a currently untranslated string to the Playback
submenu of the playback OSD menu when a playlist is active.

(cherry picked from commit 3bbf624ea685dacef29ec164cdb05a54627e62c0)