Fix playback of some streams using unusually long segment.
Al Jazeera live stream use 18s segment, it could cause the startup buffer to be exhausted before we start looking for new one. Allowing seek on live stream to allow seeking on non-completed myth transcode has added a quite severe delay as FFmpeg first tries to estimate the length of the stream by looking into the last segment. This means that for playback to start, we must retrieve the first two segments, followed by the last two.
Need to find a way to 1- make the StreamWorker seek operation lockless and 2-Find a way on how to determine when a stream is really live and one that isn't (like mythtv one).

This fix introduce some kind of regression, you can only seek within the size of the file as determined at startup.