Optimise start up time and seek time
Greatly reduce the time it takes to start playback and seek. Change the StreamWorker thread to be almost lock-less so the playback stream doesn't have to wait for the end of a download in progress for a seek order to be acknowledged.
Seek time on a local network as generated by AirVideo or Myth is almost instantaneous.

Starting time is cut by half on low-speed network ; at the most you have to wait for the download of the first two segments only.

Also, now ignore invalid streams that would have otherwise turn on the VOD mode (seen with France 24 channel)