Allow deferred evaluation when creating custom recording rules.
When creating a custom rule by choosing "Custom Edit" for a program,
allow the use of variables such as "{TITLE}" and "{STARTTIME}".  The
variables are replaced with actual values from the program when the
rule is saved.  This allows storing sample searches for future re-use
that will automatically adapt to the new program.  Press "EDIT" to
toggle between immediate evaluation (the previous and default
behavior) and deferred evaluation.

This change is primarily from used mmm1976, with trivial changes by
me.

Fixes #10734