Schedule editor flexibility and other enhancements.
Note: no theme changes are required with this commit, however, two
changes are recommended.  See below.

The primary change in this commit allows options from the schedule,
storage and post processing options screens to optionally appear on
the main screen.  This serves two purposes.  One, a theme can be
targeted at novice users by putting only the most basic options on the
main screen and requiring use of the menu to access the more advanced
options.  Two, power users can put the options they use most
frequently on the amin screen.

Other enhancements to the schedule editor in this commit include the
following.

All buttons on the main screen except those for rule type and save are
now optional.  Any applicable functionality is always available by
pressing MENU so there is no longer a need to have a button for each
child option screen.  This change is mainly to make room on the main
screen for the new, optional settings.

A new, optional setting called "newrepeat" splits the kDupsNewEpi
functionality out from the poorly overloaded "dupscope" setting.  If
the "newrepeat" setting is found in the theme, the "dupscope" setting
does not include the "Record new episodes only" option.

The INFO and UPCOMING actions now work like they do in most other
places in mythfrontend.

The PREVVIEW and NEXTVIEW actions cycle through the main screen and
all of the applicable child screens.

The Filter options screen is logically at the same level as the other
child option screens.  If the theme includes a "Filter" button on the
Schedule options screen, it should either be removed or moved to the
main screen.  If the theme cascades the child option screens, the
Filter options screen should use the same positioning as the other
child option screens.

The Metadata options screen no longer performs an automatic lookup
when the inetref is not set.

An optional save button can be on any of the child options screens.