FFmpeg adjustment.
This commit tries to minimise the changes made on the original FFmpeg libs so it is easier to sync in the future. It also fixes some errors introduced during the last resync.

Lots of MMX code were relying on mmx.h header which was removed from ffmpeg years ago. So copy it into libmythbase so we never have to bother about this files again and modify all files using that header

There are still changes in libavcodec's PPC code, but I wouldn't dare touching it.