Remove the broken "Expire in deleted order" setting.
Unfortunately, this setting has been broken since the day it was
committed.  When disabled, this setting could cause non-deleted
programs to be expired before deleted programs when the Auto
Expirer was trying to free up disk space.  The desired effect
of the setting is now already handled by default, expiring deleted
programs in a FIFO order on a particular filesystem.  The first
program deleted is the first to be expired for disk space if
necessary.