Commflag: Increase the value of MAX_BLANK_FRAMES.
For regions of blank frames between "skip" sections and "keep"
sections, the blank region is divided equally between the "skip" and
"keep" sections, subject to the MAX_BLANK_FRAMES limit.  The limit is
increased from 60 to 180 frames to give a better chance of an even
split, combined with a reasonable upper limit.