Refs #10311. Reimplements RefererenceCounter using QAtomicInt for lockless reference counting.
This also renames AddRef to IncrRef and DownRef to DecrRef. This is solely for the benefit of any plugins that might be using the reference counters in MythUI when this is extended to those objects. The MythUI reference counters start at zero instead of one so anything using those will need to remove the UpRef/AddRef immediately post creation.