Refs #10311. Add additional debugging capabilities to ReferenceCounter.
The original ReferenceCounter class would print out the QObject's objectName().
I didn't implement this when I rewrote ReferenceCounter as a general purpose
reference counter. But having a name for an object does aid in debugging, this
reintroduces that idea but with a QString passed into the constructor. By
default this is not enabled, but if you need to debug reference counting
enabling it is just a matter of uncommenting a "//#define EXTRA_DEBUG" in
the header and recompiling.