Ensure no modification to FFmpeg libavformat's URLProtocol change would break our code
FFmpeg's recently changed the structure of the URLProtocol object, which broke our code.
While this time it caused a compilation error and was easy to spot, it could have changed in such a way that it would have compiled properly and cause runtime crash.

C++ doesn't have C99 designated initialisers, so do it another way. Clean up avfringbuffer while at it