libmythtv: MHEG ignores TEXTEXIT if no onscreen to allow Esc key binding
When browsing the MHEG service it is common to use the back button to
return to the previous page.  On many RCs this key sends Esc.
If the Back button (Esc) key is bound to TEXTEXIT then it isn't possible
to exit LiveTV because MHEG grabs the key.

This patch ignores TEXTEXIT key presses when there's nothing on screen
which allows the key to be passed on to TV playback and allow normal exit.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Stuart Morgan <smorgan@mythtv.org>