Add GNU FreeFont and Droid fonts to the top-level themes directory.
FreeFont is updated to the latest (2012-05-03) version, and the full
set of fonts is included.

Droid font families are added - DroidSerif, DroidSans, and
DroidSansMono.  Note that DroidSans does not have an italic version,
so Qt will perform font substitution (likely FreeSansOblique).

The copy of DroidSans in MythCenter-wide is redundant and should be
removed.