Remove unused DBCleanup housekeeping job.
The DBCleanup job was added 4 years ago in 9c0fa01f47 as a workaround
for an issue reported in #5482.  However, the job's last run time was
never updated, meaning the job only ever ran a single time on any MythTV
system since it was added.  This seems to indicate that running it as a
daily job is unnecessary.  Furthermore, DB connection handling has been
completely rewritten to use per-thread connection pools, and both MySQL
and Qt-MySQL have changed significantly in that time, so the old
workaround for an old MySQL issue is no longer required.  And
MDBManager::PurgeIdleConnections() is now used by MDBManager's
push/popConnection() functions and MPoolThread::run(), so the
originally-intended function of the patch (cleaning up idle connections
when no longer required) is now handled--more responsively--by the DB
connection pooling code, itself.