Disable strict mode in MySQL connections.
We have a lot of code that fails in MySQL strict mode, so ensure that
we're using compatible mode for MythTV DB connections.  This will likely
fix issues such as that fixed in 96cf1cfce , which I couldn't reproduce
without enabling strict mode.

Also changes line setting time_zone to use session, instead of its
synonym, local, to make it easier to find in MySQL docs.