Fix scheduler related issues resulting from the UTC changes.
By definition kTimeslot and kWeekslot recording rules work on local
time.  Change the scheduler to convert UTC times back to local time on
the fly as needed.

IMPORTANT NOTE: This means MythTV now needs working time zone support
in MySQL.  Because time zone support is not typically configured in
MySQL by most distributions at this time, users will need to do that
themselves.  See the following URL for more information.

  http://dev.mysql.com/doc/refman/5.5/en/mysql-tzinfo-to-sql.html

In mythtv-setup and mythbackend, check the MySQL time zone support and
exit immediately if it is not working.

It's simpler to keep record.findday and findtime in local time at this
point.  Undo part of schema upgrade 1305 and update RecordingRule to
accomplish that.

Change the other, recent time zone related schema upgrades to use
CONVERT_TZ() instead of simply adding the utc_offset in minutes.  This
is so future upgraders will not have the "off by 1 hour" problem on
some historical times.