Change MythSignalingTimer to use QWaitCondition
When you do a MythSignalingTimer::stop(), the method would not return until
the next timer expiry.  This is an issue as I'm using it for a QTimer
replacement with a 5 minute timer...  Changed it to use QWaitCondition such
that it is interrupted nearly immediately when ::stop() is run.  Also, unlock
the mutex before running wait() as this avoids a deadlock.

Also changed the locking in the loggingserver to avoid more possible
deadlock situations.