OSD: Add a graphical display of audio level to cut list editor
When using the cut list editor to remove ads etc it's difficult to place
the cut precisely when the program is faded in or out.  In these cases
it's common to cut the dialogue or retain voice-overs.  It's also
impossible to create a cut list for DVB radio recordings.

This patch adds a visual graph of the audio level +/-0.5 frame around the
current paused frame.

Add a mutex to guard access by the decoder and UI threads

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Jean-Yves Avenard <jyavenard@mythtv.org>