Service Contract cleanup
First, this ads warnings to programtypes.h and recordingtypes.h
to warn those modifying the enums not to renumber things.

Second, this removes use of a libmythbase header in the recording
contract. The contract doesn't need access to any of these functions.

Third, this removes the libmythbase from the includes in the pro
and adds libmyth to the depend path so qmake will check for changes
in programtypes.h and recordingtypes.h requiring a recompile.