Keep MythUIClock time in UTC internally.
I noticed the frontend doing a lot of stat(/etc/localtime) checks.
It turned out this was because keeping the time to update in localtime
this would lead to the time not being updated for an hour when the
clocks 'spring back', but really the problem is all those unnecessary
time conversions going on.