transcode: Tolerate mpeg errors when re-encoding to I-frames around cut points
If a recording has a cutlist and is transcoded using that cutlist then
if there are MPEG decoding errors around the cut point the process of
re-coding to I frames can fail and the whole transcoding fails.

This patch makes the function MPEG2fixup::ConvertToI tolerant to
decoding errors and just skips the offending frames.

Fixes #10863

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Gavin Hurlbut <ghurlbut@mythtv.org>